+++ src/components/custom-button/customButton-styles.tsx
@@ -2,7 +2,7 @@ import {StyleSheet, TextStyle, ViewStyle} from 'react-native';
 
 import {FONTS} from '../../theme/typography';
 import {COLORS} from '../../theme/colors';
-import {SPACING} from '../../theme/theme';
+import {SPACING} from '../../theme/spacing';
 
 interface IButtonStyles {
   container: ViewStyle;
