+++ src/components/notification-card/NotificationCard-styles.tsx
@@ -2,7 +2,7 @@ import {ImageStyle, StyleSheet, TextStyle, ViewStyle} from 'react-native';
 
 import {FONTS} from '../../theme/typography';
 import {COLORS} from '../../theme/colors';
-import {SPACING} from '../../theme/theme';
+import {SPACING} from '../../theme/spacing';
 
 interface ICardStyles {
   container: ViewStyle;
