+++ src/constants/dactive-constants.tsx
@@ -17,12 +17,3 @@ export const titleColors = {
 };
 
 export const D_ACTIVE_API_URL = 'http://10.0.2.2:3000/exercises';
-
-export const LINEAR_GRADIENTS = {
-  gradient1: [
-    'rgba(255, 255, 255, 0.56)',
-    'rgba(255, 255, 255, 0.5)',
-    'rgba(255, 255, 255, 0.1)',
-  ],
-  gradient2: ['rgba(255, 255, 255, 0)', 'rgba(255, 255, 255, 0.8)'],
-};
